requirements.txt
antiorm==1.2.1
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.4
db==0.1.1
db-sqlite3==0.0.1
Django==3.0.5
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
iniconfig==1.1.1
more-itertools==8.2.0
numpy==1.19.5
packaging==20.3
pandas==1.1.3
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2019.3
python-dotenv==0.15.0
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
